dedup(1) finds and clones duplicate files
Initial commit of dedup, a utility for finding and replacing duplicate
files with cloned blocks (using `clonefile(2)`) in one more more
directory trees.

This initial commit contains the program, man page, and tests. It
handles the most common use cases:

* files that are the same but not already linking the same data or
  cloned from each other will be turned into clones of one another
* replaced files retain metadata, attributes, flags, and ACLs
* files with multiple links are not replaced but may be used as the
  source of other clones
* irregular files are ignored
* volumes are checked to ensure `clonefile(2)` is supported
* `dedup` can tell you which files are duplicated without deduplicating
  them
* options allow limiting depth, preventing traversal across filesystems,
  displaying human readable output values, suppressing progress
  display, overriding the number of threads used, and more.

Until `clonefile(2)` is supported in FreeBSD (or elsewhere), this is
strictly a macOS only utility. As such, build logic is not written
portably, CommonCrypto is used (as opposed to OpenSSL or a pluggable
abstraction), and building assumes clang, codesign, MacPorts, etc.