work around unexpected behavior
There are two edge cases that don't appear to be documented and aren't
obvious from reading through `copyfile.c` or `clonefile.c`:

* `clonefile(2)` will succeed, but the created file will not share a
  `ATTR_CMNEXT_CLONEID`
* even though `COPYFILE_DATA` is not passed to `copyfile(3)`, it will
  seemingly, compress the data using LZVN, and put the resulting data
  in a resource fork.

In the first case `getattrlist` suggests that the cloned files have no
private data. `ATTR_CMNEXT_CLONEID` changes every time the file is
cloned (even though the source doesn't change). A warning message will
be emitted when this occurs. Root cause is currently unknown.

In the second case `com.apple.decmpfs` is set and it appears that the
resource fork is set to a compressed version of the file, however, I
cannot extract the contents and while this should be transparent,
programs just see a zero byte file. This appears to happen reliably for
a certain set of files, but if those files are copied, the behavior is
not reproduced.

While the files I know of can reliably reproduce the issue, once those
files are copied (even with `cp -a` on the same volume), the behavior
no longer occurs.