Eliminate Memory Leaks, Fix Lazy SHA-256 Bug
All memory for the work queeu, metadata trees, and duplicate sets is now
correctly freed when it is done being used. Paths are also duplicated
during every metadata copy to ensure ownership is correct.

Fixed a bug where SHA-256s were only being calculated on the first two
files with the same (device, size, first, last) tuple. This resulted in
every other files with that tuple being compared aginst the empty
SHA-256 resulting in a significant number of false positives. This would
have been a catastrophic error and may be a reason to do a paranoid
comparison immediately before replacing a file.