Replace `mmap` with `FILE*` when reading first/last
There is an edge case with mmap where if the file changes size between
`fstat` and `mmap` calls, the size of the memory segment returned by
`mmap` won't match and is not queryable. If the file has gotten smaller,
attempting to read past the file size results in a `SIGBUS` error.

For reading the beginning and end of a file `mmap` was just a
convenience, so it has been replaced with `fgetc` and `ftell`.

SHA-256 calculation should do the same at some point in the future.