Fix #1, #2: Special Handling for Compressed Files
When encountering compressed files, `copyfile(3)` will (unexpectedly)
clear the `UF_COMPRESSED` flag when applying metadata from the target
file to the temporary clone if the temp clone is HFS compressed and the
target file is not.

HFS transparent compression is accomplished by compressing data and
writing it to the resource fork and then setting the `com.apple.decmpfs`
attr to details about the compression (type, original file size), and
then setting the `UF_COMPRESSION` flag on the file. `chflags` won't
allow setting `UF_COMPRESSION` on a file without the `com.apple.decmpfs`
attribute, but will allow removing it. And that is exactly what
`copyfile(3)` does when copying metadata from an uncompressed file to a
compressed file. Without `UF_COMPRESSION` set, nothing treats the file
like a compressed file and the file appears empty, but the attributes
become visible in `xattr(1)`.

This is the root cause of #2. When `UF_COMPRESSION` is cleared, the file
appears as a zero byte file.

Compressed files also cause the issue found in #1. Because compressed
data is stored in a resource fork, which is a special xattr, and not in
the data fork, there are no data blocks to clone. Metadata is written to
a new file, which also has zero blocks. Since metadata isn't shared,
there is no savings (and in the case where a compressed file is
replacing a clone, required space will increase). Since both files have
zero blocks, they don't get the same clone ID.

A race condition was also found and fixed related to lazy SHA-256
loading. If the temporary metadata was returned in one thread and then
moved to a fully populated node in another thread, the memory for that
node is deallocated. This metadata is now copied while still in the
critical section to avoid the race and reference to freed memory.